requirements.txt
streamlit==1.24.1
langchain==0.0.225
python-dotenv==0.19.2
openai==0.27.8
langchain-google-genai
google-generativeai